Fix line length and style

+++ src/SharpLearning.GradientBoost/Loss/GradientBoostAbsoluteLoss.cs
@@ -5,14 +5,14 @@ using SharpLearning.GradientBoost.GBMDecisionTree;
 namespace SharpLearning.GradientBoost.Loss
 {
     /// <summary>
-    /// Least absolute deviation (LAD) loss function. LAD gives equal equal emphasis to all observations. 
+    /// Least absolute deviation (LAD) loss function. LAD gives equal emphasis to all observations. 
     /// This makes LAD robust against outliers.
     /// http://en.wikipedia.org/wiki/Least_absolute_deviations
     /// </summary>
     public sealed class GradientBoostAbsoluteLoss : IGradientBoostLoss
     {
         /// <summary>
-        /// Least absolute deviation (LAD) loss function. LAD gives equal equal emphasis to all observations. 
+        /// Least absolute deviation (LAD) loss function. LAD gives equal emphasis to all observations. 
         /// This makes LAD robust against outliers. LAD regression is also sometimes known as robust regression. 
         /// http://en.wikipedia.org/wiki/Least_absolute_deviations
         /// </summary>
@@ -64,7 +64,7 @@ namespace SharpLearning.GradientBoost.Loss
                 }
             }
 
-            splitInfo.Cost = splitInfo.SumOfSquares - (splitInfo.Sum * splitInfo.Sum / (double)splitInfo.Samples);
+            splitInfo.Cost = splitInfo.SumOfSquares - (splitInfo.Sum * splitInfo.Sum / splitInfo.Samples);
 
             return splitInfo;
         }
@@ -95,7 +95,8 @@ namespace SharpLearning.GradientBoost.Loss
         /// <param name="predictions"></param>
         /// <param name="residuals"></param>
         /// <param name="inSample"></param>
-        public void UpdateResiduals(double[] targets, double[] predictions, double[] residuals, bool[] inSample)
+        public void UpdateResiduals(double[] targets, double[] predictions, 
+            double[] residuals, bool[] inSample)
         {
             for (int i = 0; i < residuals.Length; i++)
             {
@@ -113,19 +114,20 @@ namespace SharpLearning.GradientBoost.Loss
         /// <param name="right"></param>
         /// <param name="target"></param>
         /// <param name="residual"></param>
-        public void UpdateSplitConstants(ref GBMSplitInfo left, ref GBMSplitInfo right, double target, double residual)
+        public void UpdateSplitConstants(ref GBMSplitInfo left, ref GBMSplitInfo right, 
+            double target, double residual)
         {
             var residual2 = residual * residual;
 
             left.Samples++;
             left.Sum += residual;
             left.SumOfSquares += residual2;
-            left.Cost = left.SumOfSquares - (left.Sum * left.Sum / (double)left.Samples);
+            left.Cost = left.SumOfSquares - (left.Sum * left.Sum / left.Samples);
 
             right.Samples--;
             right.Sum -= residual;
             right.SumOfSquares -= residual2;
-            right.Cost = right.SumOfSquares - (right.Sum * right.Sum / (double)right.Samples);
+            right.Cost = right.SumOfSquares - (right.Sum * right.Sum / right.Samples);
         }
 
         /// <summary>
@@ -136,7 +138,8 @@ namespace SharpLearning.GradientBoost.Loss
         /// <param name="predictions"></param>
         /// <param name="inSample"></param>
         /// <returns></returns>
-        public double UpdatedLeafValue(double currentLeafValue, double[] targets, double[] predictions, bool[] inSample)
+        public double UpdatedLeafValue(double currentLeafValue, double[] targets, 
+            double[] predictions, bool[] inSample)
         {
             var values = new List<double>();
